Cleanup

+++ commands/staff/addrole.ts
@@ -1,10 +1,10 @@
 import * as commando from 'discord.js-commando'
 import { oneLine } from 'common-tags'
-import * as Logger from '../../util/Logger'
+import { log } from '../../util'
 import * as moment from 'moment'
 import { Message, TextChannel, Role, User, GuildChannel } from 'discord.js';
 
-module.exports = class AddRoleCommand extends commando.Command {
+export default class AddRoleCommand extends commando.Command {
   constructor(client) {
     super(client, {
       name: 'addrole',
@@ -20,7 +20,7 @@ module.exports = class AddRoleCommand extends commando.Command {
       guildOnly: true,
 
       args: [{
-        key: 'username',
+        key: 'user',
         label: 'username',
         prompt: 'Who would you like to add the role to?',
         type: 'user',
@@ -37,33 +37,37 @@ module.exports = class AddRoleCommand extends commando.Command {
     })
   }
 
-  async run(msg: commando.CommandMessage, { username, role }: { username: User, role: Role }): Promise<Message | Message[]> {
-    if (!msg.member.hasPermission("ADMINISTRATOR")) {
-      await msg.reply("Too bad, so sad. Very sad.")
+  async run(msg: commando.CommandMessage, { user, role }: { user: User, role: Role }): Promise<Message | Message[]> {
+    if (!msg.member.hasPermission('MANAGE_ROLES')) {
+      await msg.reply('You can\'t add roles.')
       return msg.delete()
     }
 
-    const user = msg.guild.member(username)
+    const member = msg.guild.member(user)
 
-    try {
-      const channel = msg.guild.channels.find((channel: GuildChannel) => channel.name === 'machobot-audit') as TextChannel
-
-      user.roles.add(role)
-
-      if (channel) {
-        channel.send(`${msg.author.username} has added ${role.name} to ${user.displayName}.`)
-      }
-
-      const time = moment().format('YYYY-MM-DD HH:mm:ss Z')
-      Logger.log(`\r\n[${time}] ${msg.author.username} has added ${role.name} to ${user.displayName}.`)
-
-      await msg.reply(`Added rank ${role.name} to ${user.displayName}.`)
-      msg.delete()
-    } catch (err) {
-      console.log(err)
-
-      await msg.reply("I couldn't complete your request, sorry 'bout that.")
+    if (msg.member.roles.highest.comparePositionTo(role) <= 0) {
+      await msg.reply('You can\'t add roles that are higher than or equal to yours.')
       return msg.delete()
     }
+
+    const channel = msg.guild.channels.find((channel: GuildChannel) => channel.name === 'machobot-audit') as TextChannel
+    const addRoleResponse = await member.roles.add(role).catch(() => {
+      return
+    })
+
+    if (!addRoleResponse) {
+      await msg.reply("I don't have permission to add that role to that user.")
+      return msg.delete()
+    }
+
+    if (channel) {
+      channel.send(`${msg.author.username} has added role ${role.name} to ${member.displayName}.`)
+    }
+
+    const time = moment().format('YYYY-MM-DD HH:mm:ss Z')
+    log(`\r\n[${time}] ${msg.author.username} has added role ${role.name} to ${member.displayName}.`)
+
+    await msg.reply(`Added role ${role.name} to ${member.displayName}.`)
+    return msg.delete()
   }
 }
