+++ commands/staff/removerole.ts
@@ -1,10 +1,10 @@
 import * as commando from 'discord.js-commando'
 import { oneLine } from 'common-tags'
-import * as Logger from '../../util/Logger'
+import { log } from '../../util'
 import * as moment from 'moment'
-import { Message, TextChannel, GuildChannel, User, Role } from 'discord.js';
+import { Message, TextChannel, GuildChannel, Role, GuildMember } from 'discord.js';
 
-module.exports = class RemoveRoleCommand extends commando.Command {
+export default class RemoveRoleCommand extends commando.Command {
   constructor(client) {
     super(client, {
       name: 'removerole',
@@ -20,10 +20,10 @@ module.exports = class RemoveRoleCommand extends commando.Command {
       guildOnly: true,
 
       args: [{
-        key: 'username',
-        label: 'username',
+        key: 'member',
+        label: 'user',
         prompt: 'Who would you like to remove the role from?',
-        type: 'user',
+        type: 'member',
         infinite: false
       },
       {
@@ -37,30 +37,35 @@ module.exports = class RemoveRoleCommand extends commando.Command {
     })
   }
 
-  async run(msg: commando.CommandMessage, { username, role }: { username: User, role: Role }): Promise<Message | Message[]> {
-    if (!msg.member.hasPermission("ADMINISTRATOR")) {
-      await msg.reply("FeelsBad(Wo)Man (it's 2018)") as Message
+  async run(msg: commando.CommandMessage, { member, role }: { member: GuildMember, role: Role }): Promise<Message | Message[]> {
+    if (!msg.member.hasPermission('MANAGE_ROLES')) {
+      await msg.reply('You can\'t remove roles.')
       return msg.delete()
     }
 
-    const user = msg.guild.member(username)
-    try {
-      const channel = msg.guild.channels.find((channel: GuildChannel) => channel.name === 'machobot-audit') as TextChannel
-
-      user.roles.remove(role)
-
-      if (channel) {
-        channel.send(`${msg.author.username} has removed ${role.name} from ${user.displayName}.`)
-      }
-
-      const time = moment().format('YYYY-MM-DD HH:mm:ss Z')
-      Logger.log(`\r\n[${time}] ${msg.author.username} has removed ${role.name} from ${user.displayName}.`)
-
-      await msg.reply(`Removed rank ${role.name} from ${user.displayName}.`) as Message
-      return msg.delete()
-    } catch (err) {
-      await msg.reply("I couldn't complete your request, sorry 'bout that.") as Message
+    if (msg.member.roles.highest.comparePositionTo(role) <= 0) {
+      await msg.reply('You can\'t remove roles that are higher than or equal to yours.')
       return msg.delete()
     }
+
+    const channel = msg.guild.channels.find((channel: GuildChannel) => channel.name === 'machobot-audit') as TextChannel
+    const removeRoleResponse = await member.roles.remove(role).catch(() => {
+      return
+    })
+
+    if (!removeRoleResponse) {
+      await msg.reply('I can\'t remove roles.')
+      return msg.delete()
+    }
+
+    if (channel) {
+      channel.send(`${msg.author.username} has removed role ${role.name} from ${member.displayName}.`)
+    }
+
+    const time = moment().format('YYYY-MM-DD HH:mm:ss Z')
+    log(`\r\n[${time}] ${msg.author.username} has removed role ${role.name} from ${member.displayName}.`)
+
+    await msg.reply(`Removed role ${role.name} from ${member.displayName}.`)
+    return msg.delete()
   }
 }
